requirements.txt
torch
torchvision
pillow==9.0.1
paddlepaddle
paddlehub
gradio